migrate some methods to BaseActivity schema
